refactor, batch-embed with semaphore
